requirements.txt
fastapi
uvicorn[standard]
transformers
torch
python-multipart
python-dotenv